Partial function constructor to cut wind velocity (or others) into factor classes

Partial function constructor to cut wind velocity (or others) into factor classes

Usage

cut_ws.fun(
  binwidth = 1,
  ws_max = NA,
  squish = TRUE,
  right = TRUE,
  reverse = FALSE,
  calm = NA
)

Arguments

binwidth

width of the bins

ws_max

cut off wind speed at this maximum

squish

If TRUE wind velocities greater than will be include as additional level ">ws_max"

right

logical, indicating if the intervals should be closed on the right (and open on the left) or vice versa.

reverse

reverse order of result. This is sometimes useful when plotting a factor.

calm

threshold for calm wind situation. All windspeed below this limit will be assigned the the factor level "calm".

Value

a partial cut_ws() function with ws as sole argument

See also