Summarise z values over binned wind data.

Binning is done by StatWind, so input data to stat_summary_wind() should be original unbinned data. Depending on the groups argument, binning is either done 2-dimensional over cartesian u and v wind vectors (calculated from input data; then, stat_summary_wind() yields results similar to openair::polarPlot() or, 1-dimensional over wind direction or wind velocity bins, respectively.

stat_summary_wind(data = NULL, mapping = NULL, fun = "mean",
  fun.args = list(), nmin = 1, wd_cutfun = cut_wd.fun(binwidth = 45),
  wd_offset = 0, ws_cutfun = cut_ws.fun(binwidth = 1, ws_max = NA),
  groups = NULL, layer_args = list(geom = "polygon"),
  param_args = list())

Arguments

data
The data to be displayed in this layer. requires input data including at least three columns carrying information regarding: * wind direction (in °) * wind velocity * z-values (e.g. air pollutant concentration)

mapping
ggplot2 mapping, e.g. aes(wd = wd, ws = ws, z = NOx); requires wd, ws, z

fun
function or list of functions for summary.

fun.args
a list of extra arguments to pass to fun.

nmin
numeric, minimum number of values for fun, if n < nmin: NA is returned

wd_cutfun
function, cut function for wind direction (to create bins)

wd_offset
numeric, offset for wind_direction (in °) if groups == "wd"; bins are then calculated over (wd + wd_offset) %% 360

ws_cutfun
function, cut function for wind speed

groups
character string, can be NULL, "wd", ...

layer_args
named list, passed on to ggplot2::layer()

param_args
named list, passed on to ggplot2::layer() as argument params after combining with the other arguments

...
other arguments passed on to ggplot2::layer() as params = list(...).

geom
The geometric object to use display the data.

Value

ggplot2::layer()