Summarise z values over binned wind data.

Input data should be original unbinned data. 1-dimensional binning and calculating summary statistics over wind direction and/or wind velocity bins, respectively. NA values in z and in ws, wd (after cutting) will be silently removed before applying functions

summary_wind(data, ws, wd, z, groupings = groups(), fun = "mean",
  fun.args = list(), nmin = 3, wd_cutfun = cut_wd.fun(binwidth = 45),
  ws_cutfun = cut_ws.fun(binwidth = 1))

Arguments

data
a data.frame or tibble containing the data (wide format)

ws
symbol giving the wind velocity parameter name (wind velocity preferably in m/s)

wd
symbol giving the wind direction parameter name in degrees

z
symbol giving the parameter name to be summarised

groupings
additional groupings. Use helper groups() to create

fun
function or list of functions for summary.

fun.args
a list of extra arguments passed on to fun.

nmin
numeric, minimum number of values for fun, if n < nmin: NA is returned

wd_cutfun
function, cut function for wind direction (to create bins). See wd_cut.fun() for options.

ws_cutfun
function, cut function for wind speed. See ws_cut.fun() for examples

Value

a tibble with summarised data

Computed variables

The names of the columns stay the same as in data (for the arguments "ws", "wd" and "z").

binned column ws

binned column wd

factor column stat containing the name of the summarize function as level

column z with the result of the summarize function

n count of occurrences for the corresponding bin

freq frequency of the corresponding bin

Examples

fn <- rOstluft.data::f("Zch_Stampfenbachstrasse_2010-2014.csv")
data <- rOstluft::read_airmo_csv(fn)
data <- rOstluft::rolf_to_openair(data)
data <- dplyr::mutate(data, year = lubridate::year(date))

summary_wind(data, "ws", "wd", "NO2")

#> # A tibble: 57 x 6
#>    wd    ws        n     freq stat    NO2
#>    <fct> <fct> <int>    <dbl> <fct> <dbl>
#>  1 0     [0,1]  3230 0.0372   mean   43.6
#>  2 0     (1,2]  5764 0.0664   mean   39.4
#>  3 0     (2,3]  2017 0.0232   mean   36.7
#>  4 0     (3,4]   257 0.00296  mean   32.7
#>  5 0     (4,5]    22 0.000253 mean   19.9
#>  6 45    [0,1]  3894 0.0448   mean   44.6
#>  7 45    (1,2] 10172 0.117    mean   41.3
#>  8 45    (2,3]  6036 0.0695   mean   34.3
#>  9 45    (3,4]  1871 0.0215   mean   28.0
#> 10 45    (4,5]   417 0.00480  mean   24.5
#> # ... with 47 more rows

# multiple stats: Pass function, by name, reference, as function or one sided formula
q95 <- function(x) stats::quantile(x, probs = 0.95)

funs <- list(
  "mean",
  "median",
  "q95_1" = q95,
  "q95_2" = function(x) stats::quantile(x, probs = 0.95),
  "q95_3" = ~ stats::quantile(., probs = 0.95)
)

res <- summary_wind(data, "ws", "wd", "NO2", fun = funs)
res

#> # A tibble: 285 x 6
#>    wd    ws        n     freq stat    NO2
#>    <fct> <fct> <int>    <dbl> <fct> <dbl>
#>  1 0     [0,1]  3230 0.0372   mean   43.6
#>  2 0     (1,2]  5764 0.0664   mean   39.4
#>  3 0     (2,3]  2017 0.0232   mean   36.7
#>  4 0     (3,4]   257 0.00296  mean   32.7
#>  5 0     (4,5]    22 0.000253 mean   19.9
#>  6 45    [0,1]  3894 0.0448   mean   44.6
#>  7 45    (1,2] 10172 0.117    mean   41.3
#>  8 45    (2,3]  6036 0.0695   mean   34.3
#>  9 45    (3,4]  1871 0.0215   mean   28.0
#> 10 45    (4,5]   417 0.00480  mean   24.5
#> # ... with 275 more rows

# and q95, q95_1, q95_2, q95_3 alle have the same value
tidyr::spread(res, "stat", "NO2") %>%
  dplyr::select(q95_1, q95_2, q95_3)

#> # A tibble: 57 x 3
#>    q95_1 q95_2 q95_3
#>    <dbl> <dbl> <dbl>
#>  1  79.8  79.8  79.8
#>  2  74.2  74.2  74.2
#>  3  67.1  67.1  67.1
#>  4  79.6  79.6  79.6
#>  5  33.1  33.1  33.1
#>  6  81.6  81.6  81.6
#>  7  77.1  77.1  77.1
#>  8  65.4  65.4  65.4
#>  9  52.2  52.2  52.2
#> 10  45.4  45.4  45.4
#> # ... with 47 more rows

# is for some reason fun.args used with multiple functions, use ... to catch
# superfluous arguments:
funs <- list(
  "q95" = function(x, ...) stats::quantile(x, probs = 0.95),
  "mean"
)
summary_wind(data, "ws", "wd", "NO2", fun = funs, fun.args = list(na.rm = TRUE))

#> # A tibble: 114 x 6
#>    wd    ws        n     freq stat    NO2
#>    <fct> <fct> <int>    <dbl> <fct> <dbl>
#>  1 0     [0,1]  3230 0.0372   q95    79.8
#>  2 0     (1,2]  5764 0.0664   q95    74.2
#>  3 0     (2,3]  2017 0.0232   q95    67.1
#>  4 0     (3,4]   257 0.00296  q95    79.6
#>  5 0     (4,5]    22 0.000253 q95    33.1
#>  6 45    [0,1]  3894 0.0448   q95    81.6
#>  7 45    (1,2] 10172 0.117    q95    77.1
#>  8 45    (2,3]  6036 0.0695   q95    65.4
#>  9 45    (3,4]  1871 0.0215   q95    52.2
#> 10 45    (4,5]   417 0.00480  q95    45.4
#> # ... with 104 more rows

# more wd classes, less ws classes and squish ws
summary_wind(data, "ws", "wd", "NO",
             wd_cutfun = cut_wd.fun(binwidth = 22.5),
             ws_cutfun = cut_ws.fun(binwidth = 2, ws_max = 6))

#> # A tibble: 54 x 6
#>    wd    ws        n      freq stat     NO
#>    <fct> <fct> <int>     <dbl> <fct> <dbl>
#>  1 0     [0,2]  4225 0.0486    mean  27.0 
#>  2 0     (2,4]   931 0.0107    mean  22.6 
#>  3 0     (4,6]    17 0.000196  mean   6.06
#>  4 22.5  [0,2]  8338 0.0960    mean  34.5 
#>  5 22.5  (2,4]  3290 0.0379    mean  26.4 
#>  6 22.5  (4,6]    22 0.000253  mean  12.8 
#>  7 45    [0,2]  7623 0.0878    mean  28.4 
#>  8 45    (2,4]  5228 0.0602    mean  17.6 
#>  9 45    (4,6]   446 0.00513   mean   8.89
#> 10 45    >6        4 0.0000461 mean   3.19
#> # ... with 44 more rows

# only one ws class ()
summary_wind(data, "ws", "wd", "NO",
             ws_cutfun = cut_number.fun(1))

#> # A tibble: 8 x 6
#>   wd    ws                 n   freq stat     NO
#>   <fct> <fct>          <int>  <dbl> <fct> <dbl>
#> 1 0     [0.00477,9.73] 11292 0.130  mean   27.8
#> 2 45    [0.00477,9.73] 22480 0.259  mean   25.9
#> 3 90    [0.00477,9.73]  2564 0.0295 mean   21.1
#> 4 135   [0.00477,9.73]  3422 0.0394 mean   16.7
#> 5 180   [0.00477,9.73] 13007 0.150  mean   18.3
#> 6 225   [0.00477,9.73] 12088 0.139  mean   13.5
#> 7 270   [0.00477,9.73] 13197 0.152  mean   14.8
#> 8 315   [0.00477,9.73]  8809 0.101  mean   15.0

# additional grouping with strings, symbols or named expressions
summary_wind(data, ws, wd, NO2, group = groups("site", year, wday = lubridate::wday(date)))

#> Warning: Assigning non-quosure objects to quosure lists is deprecated as of rlang 0.3.0.
#> Please coerce to a bare list beforehand with `as.list()`
#> This warning is displayed once per session.

#> # A tibble: 1,284 x 9
#>    wd    ws    site                     year  wday     n   freq stat    NO2
#>    <fct> <fct> <fct>                   <dbl> <dbl> <int>  <dbl> <fct> <dbl>
#>  1 0     [0,1] Zch_Stampfenbachstrasse  2010     1    76 0.0304 mean   37.1
#>  2 0     [0,1] Zch_Stampfenbachstrasse  2010     2    86 0.0350 mean   47.2
#>  3 0     [0,1] Zch_Stampfenbachstrasse  2010     3    98 0.0398 mean   48.4
#>  4 0     [0,1] Zch_Stampfenbachstrasse  2010     4   102 0.0414 mean   49.5
#>  5 0     [0,1] Zch_Stampfenbachstrasse  2010     5    82 0.0332 mean   46.3
#>  6 0     [0,1] Zch_Stampfenbachstrasse  2010     6    85 0.0343 mean   49.9
#>  7 0     [0,1] Zch_Stampfenbachstrasse  2010     7    89 0.0357 mean   40.8
#>  8 0     [0,1] Zch_Stampfenbachstrasse  2011     1    97 0.0396 mean   30.9
#>  9 0     [0,1] Zch_Stampfenbachstrasse  2011     2    88 0.0357 mean   44.6
#> 10 0     [0,1] Zch_Stampfenbachstrasse  2011     3    88 0.0356 mean   46.9
#> # ... with 1,274 more rows

# how often comes which concentration from one direction
summary_wind(data, ws, wd, NO2,
             group = groups(NO2_class = ggplot2::cut_number(NO2, 5)),
             ws_cutfun = cut_number.fun(1))

#> # A tibble: 40 x 7
#>    wd    ws             NO2_class        n   freq stat    NO2
#>    <fct> <fct>          <fct>        <int>  <dbl> <fct> <dbl>
#>  1 0     [0.00477,9.73] [0.066,15.9]   990 0.0570 mean   12.3
#>  2 0     [0.00477,9.73] (15.9,25.6]   1992 0.115  mean   20.9
#>  3 0     [0.00477,9.73] (25.6,36.5]   2551 0.147  mean   31.1
#>  4 0     [0.00477,9.73] (36.5,51.2]   2751 0.158  mean   43.4
#>  5 0     [0.00477,9.73] (51.2,148]    3008 0.173  mean   65.8
#>  6 45    [0.00477,9.73] [0.066,15.9]  2177 0.125  mean   12.2
#>  7 45    [0.00477,9.73] (15.9,25.6]   4142 0.238  mean   21.0
#>  8 45    [0.00477,9.73] (25.6,36.5]   5510 0.317  mean   31.0
#>  9 45    [0.00477,9.73] (36.5,51.2]   5451 0.314  mean   43.3
#> 10 45    [0.00477,9.73] (51.2,148]    5200 0.299  mean   66.4
#> # ... with 30 more rows

# the same but we use ws as pollutant
summary_wind(data, NO2, wd, NO2, ws_cutfun = cut_number.fun(5))

#> # A tibble: 40 x 6
#>    wd    NO2              n   freq stat  NO2.stat
#>    <fct> <fct>        <int>  <dbl> <fct>    <dbl>
#>  1 0     [0.066,15.9]   991 0.0114 mean      12.3
#>  2 0     (15.9,25.6]   1991 0.0229 mean      20.9
#>  3 0     (25.6,36.5]   2551 0.0294 mean      31.1
#>  4 0     (36.5,51.2]   2751 0.0317 mean      43.4
#>  5 0     (51.2,148]    3008 0.0346 mean      65.8
#>  6 45    [0.066,15.9]  2177 0.0251 mean      12.2
#>  7 45    (15.9,25.6]   4145 0.0477 mean      21.0
#>  8 45    (25.6,36.5]   5507 0.0634 mean      31.0
#>  9 45    (36.5,51.2]   5450 0.0627 mean      43.3
#> 10 45    (51.2,148]    5201 0.0599 mean      66.4
#> # ... with 30 more rows