Summarise y values over binned wind data, split into u and v components.

Input data should be original unbinned data including wind direction and wind velocity; binning is done 2-dimensional over cartesian u and v wind vectors

summary_wind_2d(data, ws, wd, z, groups = NULL, fun = "mean",
  fun.args = list(), nmin = 3, ws_max = NA, bins = 100,
  smooth = TRUE, k = 100, extrapolate = TRUE, dist = 0.1)

Arguments

data
a data.frame or tibble containing the data (wide format). requires input data including at least three columns carrying information regarding:

wind direction (in °)

wind velocity

z-values (e.g. air pollutant concentration)

ws
string giving the wind velocity parameter name (wind velocity preferably in m/s)

wd
string giving the wind direction parameter name in degrees

z
string giving the parameter name to be summarised

groups
can be NULL, c("u", "v"), "wd_class", "ws_class", ...

fun
function or list of functions for summary.

fun.args
a list of extra arguments to pass to fun.

nmin
numeric, minimum number of values for fun, if n < nmin: NA is returned

ws_max
numeric or NA, maximum wind velocity for binning: above ws_max, z is set NA

bins
numeric, number of bins over the range of values if !groups %in% c("u", "v")

smooth
TRUE/FALSE, applies if groups = c("u", "v"); should smoothing of summary results should be performed using fit_gam_surface()?

k
numeric, applies if smooth = TRUE; degree of smoothing in smooth term in fit_gam_surface()

extrapolate
TRUE/FALSE, applies if smooth = TRUE; fit_gam_surface() returns extrapolated (predicted) values for u, v coordinates that otherwise would have have NA for summarised z if extrapolate = TRUE, those values are returned (to a certain degree depending on the value of dist)

dist
numeric, fraction of 1, applies if smooth = TRUE and extrapolate = TRUE; maximum distance to next coordinate-pair at which the result of fit_gam_surface(z) should be returned

Value

a tibble with summarised data along u and v wind vectors

Computed variables

a tibble is returned, binned over u and v, with variables:

wd: wind direction corresponding to midpoint value of u and v

ws: wind velocity corresponding to midpoint value of u and v

u: bins over u (from input wd and ws)

v: bins over v (from input wd and ws)

z: result from fun(z, ...)