Summarise z values over binned wind data.

Input data should be original unbinned data. 1-dimensional binning and calculating summary statistics over wind direction and/or wind velocity bins, respectively. NA values in z will be silently removed before applying functions

summary_wind(data, ws, wd, z, groups = c(), fun = "mean",
  fun.args = list(), nmin = 3, wd_cutfun = cut_wd.fun(binwidth = 45),
  ws_cutfun = cut_ws.fun(binwidth = 1))

Arguments

data
a data.frame or tibble containing the data (wide format)

ws
symbol giving the wind velocity parameter name (wind velocity preferably in m/s)

wd
symbol giving the wind direction parameter name in degrees

z
symbol giving the parameter name to be summarised

groups
one or more character string for additional grouping

fun
function or list of functions for summary.

fun.args
a list of extra arguments passed on to fun.

nmin
numeric, minimum number of values for fun, if n < nmin: NA is returned

wd_cutfun
function, cut function for wind direction (to create bins). See wd_cut.fun() for options.

ws_cutfun
function, cut function for wind speed. See ws_cut.fun() for examples

Value

a tibble with summarised data

Computed variables

The names of the columns stay the same as in data (for the arguments "ws", "wd" and "z").

binned column ws

binned column wd

factor column stat containing the name of the summarize function as level

column z with the result of the summarize function

n count of occurrences for the corresponding bin

freq frequency of the corresponding bin

Examples

fn <- rOstluft.data::f("Zch_Stampfenbachstrasse_2010-2014.csv")
data <- rOstluft::read_airmo_csv(fn)
data <- rOstluft::rolf_to_openair(data)
data <- dplyr::mutate(data, year = lubridate::year(date))

summary_wind(data, "ws", "wd", "NO2")

#> # A tibble: 57 x 6
#>    wd    ws        n stat    NO2     freq
#>    <fct> <fct> <int> <fct> <dbl>    <dbl>
#>  1 0     [0,1]  3230 mean   43.6 0.0372  
#>  2 0     (1,2]  5764 mean   39.4 0.0664  
#>  3 0     (2,3]  2017 mean   36.7 0.0232  
#>  4 0     (3,4]   257 mean   32.7 0.00296 
#>  5 0     (4,5]    22 mean   19.9 0.000253
#>  6 45    [0,1]  3894 mean   44.6 0.0448  
#>  7 45    (1,2] 10172 mean   41.3 0.117   
#>  8 45    (2,3]  6036 mean   34.3 0.0695  
#>  9 45    (3,4]  1871 mean   28.0 0.0215  
#> 10 45    (4,5]   417 mean   24.5 0.00480 
#> # ... with 47 more rows

# works with symbols too and additional grouping
summary_wind(data, ws, wd, NO2, groups = c("site", "year"))

#> # A tibble: 235 x 8
#>    wd    ws    site                     year     n stat    NO2    freq
#>    <fct> <fct> <fct>                   <dbl> <int> <fct> <dbl>   <dbl>
#>  1 0     [0,1] Zch_Stampfenbachstrasse  2010   618 mean   45.9 0.00711
#>  2 0     [0,1] Zch_Stampfenbachstrasse  2011   637 mean   43.8 0.00733
#>  3 0     [0,1] Zch_Stampfenbachstrasse  2012   691 mean   41.9 0.00796
#>  4 0     [0,1] Zch_Stampfenbachstrasse  2013   632 mean   45.9 0.00728
#>  5 0     [0,1] Zch_Stampfenbachstrasse  2014   652 mean   40.7 0.00751
#>  6 0     (1,2] Zch_Stampfenbachstrasse  2010  1177 mean   42.6 0.0136 
#>  7 0     (1,2] Zch_Stampfenbachstrasse  2011  1228 mean   39.7 0.0141 
#>  8 0     (1,2] Zch_Stampfenbachstrasse  2012  1107 mean   38.0 0.0127 
#>  9 0     (1,2] Zch_Stampfenbachstrasse  2013  1164 mean   40.2 0.0134 
#> 10 0     (1,2] Zch_Stampfenbachstrasse  2014  1088 mean   36.1 0.0125 
#> # ... with 225 more rows

# multiple stats: Pass function, by name, reference, as function or one sided formula
q95 <- function(x) stats::quantile(x, probs = 0.95)

funs <- list(
  "mean",
  "median",
  "q95_1" = q95,
  "q95_2" = function(x) stats::quantile(x, probs = 0.95),
  "q95_3" = ~ stats::quantile(., probs = 0.95)
)

res <- summary_wind(data, "ws", "wd", "NO2", fun = funs)
res

#> # A tibble: 285 x 6
#>    wd    ws        n stat    NO2      freq
#>    <fct> <fct> <int> <fct> <dbl>     <dbl>
#>  1 0     [0,1]  3230 mean   43.6 0.00744  
#>  2 0     (1,2]  5764 mean   39.4 0.0133   
#>  3 0     (2,3]  2017 mean   36.7 0.00464  
#>  4 0     (3,4]   257 mean   32.7 0.000592 
#>  5 0     (4,5]    22 mean   19.9 0.0000507
#>  6 45    [0,1]  3894 mean   44.6 0.00897  
#>  7 45    (1,2] 10172 mean   41.3 0.0234   
#>  8 45    (2,3]  6036 mean   34.3 0.0139   
#>  9 45    (3,4]  1871 mean   28.0 0.00431  
#> 10 45    (4,5]   417 mean   24.5 0.000960 
#> # ... with 275 more rows

# and q95, q95_1, q95_2, q95_3 alle have the same value
tidyr::spread(res, "stat", "NO2") %>%
  dplyr::select(q95_1, q95_2, q95_3)

#> # A tibble: 57 x 3
#>    q95_1 q95_2 q95_3
#>    <dbl> <dbl> <dbl>
#>  1  79.8  79.8  79.8
#>  2  74.2  74.2  74.2
#>  3  67.1  67.1  67.1
#>  4  79.6  79.6  79.6
#>  5  33.1  33.1  33.1
#>  6  81.6  81.6  81.6
#>  7  77.1  77.1  77.1
#>  8  65.4  65.4  65.4
#>  9  52.2  52.2  52.2
#> 10  45.4  45.4  45.4
#> # ... with 47 more rows

# is for some reason fun.args used with multiple functions, use ... to catch
# superfluous arguments:
funs <- list(
  "q95" = function(x, ...) stats::quantile(x, probs = 0.95),
  "mean"
)
summary_wind(data, "ws", "wd", "NO2", fun = funs, fun.args = list(na.rm = TRUE))

#> # A tibble: 114 x 6
#>    wd    ws        n stat    NO2     freq
#>    <fct> <fct> <int> <fct> <dbl>    <dbl>
#>  1 0     [0,1]  3230 q95    79.8 0.0186  
#>  2 0     (1,2]  5764 q95    74.2 0.0332  
#>  3 0     (2,3]  2017 q95    67.1 0.0116  
#>  4 0     (3,4]   257 q95    79.6 0.00148 
#>  5 0     (4,5]    22 q95    33.1 0.000127
#>  6 45    [0,1]  3894 q95    81.6 0.0224  
#>  7 45    (1,2] 10172 q95    77.1 0.0586  
#>  8 45    (2,3]  6036 q95    65.4 0.0347  
#>  9 45    (3,4]  1871 q95    52.2 0.0108  
#> 10 45    (4,5]   417 q95    45.4 0.00240 
#> # ... with 104 more rows

# more wd classes, less ws classes and squish ws
summary_wind(data, "ws", "wd", "NO",
             wd_cutfun = cut_wd.fun(binwidth = 22.5),
             ws_cutfun = cut_ws.fun(binwidth = 2, ws_max = 6))

#> # A tibble: 54 x 6
#>    wd    ws        n stat     NO      freq
#>    <fct> <fct> <int> <fct> <dbl>     <dbl>
#>  1 0     [0,2]  4225 mean  27.0  0.0486   
#>  2 0     (2,4]   931 mean  22.6  0.0107   
#>  3 0     (4,6]    17 mean   6.06 0.000196 
#>  4 22.5  [0,2]  8338 mean  34.5  0.0960   
#>  5 22.5  (2,4]  3290 mean  26.4  0.0379   
#>  6 22.5  (4,6]    22 mean  12.8  0.000253 
#>  7 45    [0,2]  7623 mean  28.4  0.0878   
#>  8 45    (2,4]  5228 mean  17.6  0.0602   
#>  9 45    (4,6]   446 mean   8.89 0.00513  
#> 10 45    >6        4 mean   3.19 0.0000461
#> # ... with 44 more rows

# only one ws class ()
summary_wind(data, "ws", "wd", "NO",
             ws_cutfun = function(x) factor("ws"))

#> # A tibble: 8 x 6
#>   wd    ws        n stat     NO   freq
#>   <fct> <fct> <int> <fct> <dbl>  <dbl>
#> 1 0     ws    11292 mean   27.8 0.130 
#> 2 45    ws    22480 mean   25.9 0.259 
#> 3 90    ws     2564 mean   21.1 0.0295
#> 4 135   ws     3422 mean   16.7 0.0394
#> 5 180   ws    13007 mean   18.3 0.150 
#> 6 225   ws    12088 mean   13.5 0.139 
#> 7 270   ws    13197 mean   14.8 0.152 
#> 8 315   ws     8809 mean   15.0 0.101